separate todo list function as a hook
